requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
aiohttp==3.10.5
shazamio==0.4.0.1
python-dotenv==1.0.1
yt-dlp>=2024.10.7
